Update README.md
Removed the Base64<-->ASCII functions and scripts due to the inefficiencies of the syncing process using these commands to read from and write to the XML files stored in container fields.

To write the XML data to a file the database now uses the [Export Field Contents] script step and then imports the newly created XML file into a container field.

To read the data from the imported XML file the database now uses a few tricks to access the contents.  The file is exported using the [Export Field Contents] script step.  Then the contents of the XML file are imported into the XmlImport table line by line.  The XmlImport records are then combined back into a single field/variable to process the sync updates.